Tests/LibWeb/Text/expected/css/inserted-subtree-styles-are-computed.txt

leaf after subtree insert: rgb(0, 128, 0)
direct leaf after insert: rgb(0, 0, 255)
